requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
pydantic==1.10.11
transformers==4.34.0
torch==2.4.1
python-dotenv==1.0.0
